+++ Tophat-Railway/src/application/TrackModel/TrackModelCtrl.java
@@ -19,7 +19,6 @@ import javafx.scene.control.Label;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.shape.Circle;
 import javafx.scene.shape.Line;
-import javafx.scene.shape.LineBuilder;
 import javafx.fxml.Initializable;
 
 public class TrackModelCtrl implements Initializable {
