+++ Tophat-Railway/src/application/TrainModel/TrainModelInterface.java
@@ -1,0 +1,14 @@
+package application.TrainModel;
+
+import application.TrackModel.TrackBlock;
+
+public interface TrainModelInterface {
+
+    TrainModel getTrain(String trainID);
+
+    TrainModel createTrain(String trainID);
+
+    boolean dispatchTrain(String trainID);
+
+    void makeTrain(int i, int i1, int i2, TrackBlock trackBlock, TrackBlock trackBlock1);
+}
