+++ Tophat-Railway/src/application/TrainModel/TrainModelMain.java
@@ -1,10 +1,14 @@
 package application.TrainModel;
+/**
+ * This is test code for the train model.
+ */
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
+
 
 public class TrainModelMain extends Application {
