+++ Tophat-Railway/src/application/TrainModel/TrainModelSingleton.java
@@ -1,52 +1,67 @@
 package application.TrainModel;
 
-import application.CTC.CTCSingleton;
-import application.MBO.MBOSingleton;
-import application.TrackController.TrackControllerSingleton;
+import application.TrackModel.TrackBlock;
 import application.TrackModel.TrackModelSingleton;
-import application.TrainController.TrainControllerSingleton;
 
-public class TrainModelSingleton {
+import java.util.Collection;
+import java.util.HashMap;
 
-	// Singleton Functions (NO TOUCHY!!)
-	private static TrainModelSingleton instance = null;
+public class TrainModelSingleton implements TrainModelInterface {
+    private static TrainModelSingleton ourInstance = new TrainModelSingleton();
 
-	private TrainModelSingleton() {
-	}
+    public static TrainModelSingleton getInstance() {
+        return ourInstance;
+    }
 
-	public static TrainModelSingleton getInstance() {
-		if (instance == null) {
-			instance = new TrainModelSingleton();
-		}
+    private HashMap<String, TrainModel> trainModelHashMap;
 
-		return instance;
-	}
 
-	// =====================================
+    private TrainModelSingleton() {
+        trainModelHashMap = new HashMap<>();
+    }
 
-	// NOTE: Put your data objects here
-	private int count = 0;
+    public TrainModel getTrain(String trainID){
+        return trainModelHashMap.get(trainID);
+    }
 
-	// NOTE: Put some functions here
-	public void increment() {
-		count++;
-	}
+    @Override
+    public TrainModel createTrain(String trainID) {
+        TrainModel train = new TrainModel(trainID);
+        trainModelHashMap.put(trainID, train);
+        return train;
+    }
 
-	public int getCount() {
-		return count;
-	}
+    @Override
+    public boolean dispatchTrain(String trainID) {
+        return trainModelHashMap.put(trainID, new TrainModel(trainID)) != null;
+    }
 
-	// NOTE: Singleton Connections (Put changes reads, gets, sets that you want to
-	// occur here)
-	// WARNING: This Only changes the singleton, not your UI. UI updates occur in
-	// your UI controller
-	public void update() {
-		// Example: get the count from a singleton and replace yours with the largest
-		TrackModelSingleton tckModSin = TrackModelSingleton.getInstance();
-		int t_count = tckModSin.getCount();
-		if (count < t_count)
-			count = t_count;
+    @Override
+    public void makeTrain(int i, int i1, int i2, TrackBlock trackBlock, TrackBlock trackBlock1) {
 
-	}
+    }
 
+    Collection<TrainModel> getTrains() {
+        return trainModelHashMap.values();
+    }
+
+
+
+    //====================================================
+
+    // NOTE: Singleton Connections (Put changes reads, gets, sets that you want to
+    // occur here)
+    // WARNING: This Only changes the singleton, not your UI. UI updates occur in
+    // your UI controller
+    public void update() {
+        // Example: get the count from a singleton and replace yours with the largest
+        TrackModelSingleton tckModSin = TrackModelSingleton.getInstance();
+
+
+    }
+
+    public int getCount() {
+        System.err.println("Remove this call.");
+        return 1;
+    }
 }
